+++ core/src/org/noses/games/homedefense/game/ClockTickHandler.java
@@ -2,5 +2,5 @@ package org.noses.games.homedefense.game;
 
 public interface ClockTickHandler {
 
-    public void clockTick(float delta);
+    public void clockTick(double delta);
 }
